+++ src/Renderer/mediainfo.h
@@ -164,6 +164,8 @@ public:
     void setCrop(int width, int height, int id = -1, bool silent = false);
     void setCropUseSize(bool useSize, int id = -1, bool silent = false);
     void setLut(QString lut, int id = -1, bool silent = false);
+    void setDeinterlace(bool deinterlace, int id = -1, bool silent = false);
+    void setDeinterlaceParity(MediaUtils::DeinterlaceParity parity, int id = -1, bool silent = false);
     //audio
     void setSamplingRate(int value, int id = -1, bool silent = false);
     void setChannels(QString value, int id = -1, bool silent = false);