+++ src/UI/outputwidget.h
@@ -9,6 +9,7 @@
 #include "UI/Blocks/blockbasewidget.h"
 #include "UI/Blocks/blockcrop.h"
 #include "UI/Blocks/blocklut.h"
+#include "UI/Blocks/blockdeinterlace.h"
 #include "UI/Blocks/blockresize.h"
 #include "UI/Blocks/blockframerate.h"
 #include "UI/Blocks/blockvideocodec.h"
@@ -88,6 +89,7 @@ private:
     BlockBaseWidget *addBlock(BlockContentWidget *content, QAction *action, QString icon="");
     BlockBaseWidget *blockCrop;
     BlockBaseWidget *blockLut;
+    BlockBaseWidget *blockDeinterlace;
     BlockBaseWidget *blockResize;
     BlockBaseWidget *blockFrameRate;
     BlockBaseWidget *blockVideoCodec;