+++ src-docs/docs/changelog.md
@@ -13,6 +13,7 @@ This is the list of what has changed since the first public version of *DuFFmpeg
     - Added the `--minimize` option.
 - Ability to **Crop** videos.
 - Ability to [**Apply a 1D or 3D LUT**](lut.md).
+- Ability to Deinterlace video.
 
 ### Improvements